Honor new rust 1.95 clippy lints
Rust 1.95's clippy fires additional lints that CI's -D warnings
treats as errors. Fixed each at source rather than suppressing:

- collapsible_match (11x) in src/compiler/bytecode.rs's
  disassemble_lines: convert `match Pat => { if cond { body } }`
  to `match Pat if cond => { body }` for each instruction arm.

- collapsible_match in src/lir/types.rs register_reaches_non_int:
  merged `if *dst == r` with `if seen_slots.insert(*slot)` into a
  single guard.

- collapsible_match in src/syntax/expand/syntaxcase.rs's pattern-
  variable uniqueness check: replaced
  `Pattern::Variable(name) => { if !seen.insert(...) { ... } }`
  with a guard arm + catch-all arm.

- unnecessary_sort_by in src/rewrite/edit.rs: replaced
  `sort_by(|a, b| b.offset.cmp(&a.offset))` with
  `sort_by_key(|e| Reverse(e.byte_offset))`.

- new_without_default in src/mlir/cache.rs: added Default impl
  for MlirCache forwarding to Self::new.

- needless_borrow in src/mlir/spirv.rs: dropped an extra `&` on
  PassManager::new(context).

- too_many_arguments in src/mlir/spirv.rs emit_if_return: bundled
  the six scalar parameters (three block indices + three string
  contexts) into an IfReturn struct.

- unnecessary_cast in src/primitives/meta.rs prim_git: dropped the
  redundant `as i64` on `.as_int().unwrap_or(256)`, which already
  returns i64.

Local `cargo clippy --workspace --all-targets --all-features -- -D warnings`
is clean with rust 1.95.0.

+++ src/primitives/meta.rs
@@ -550,7 +550,7 @@ pub(crate) fn prim_git(args: &[Value]) -> (SignalBits, Value) {
             );
         }
         let wg_size = if args.len() == 2 {
-            args[1].as_int().unwrap_or(256) as i64
+            args[1].as_int().unwrap_or(256)
         } else {
             256
         };
